Refactor groups package into root
