+++ README.md
@@ -15,3 +15,9 @@ pip install rolumns
 ## Contributing
 
 Please read [CONTRIBUTING.md](https://github.com/cariad/rolumns/blob/main/CONTRIBUTING.md) before raising an issue or submitting a pull request. Thank you!
+
+## Author
+
+Hello! 👋 I'm **Cariad Eccleston**, and I'm a freelance Amazon Web Services architect, DevOps evangelist, and infrastructure and backend engineer by the beach in the United Kingdom.
+
+You can find me at [cariad.earth](https://cariad.earth), [github/cariad](https://github.com/cariad) and [linkedin/cariad](https://linkedin.com/in/cariad).
