+++ docs/source/index.rst
@@ -30,3 +30,10 @@ Rolumns is released under the MIT Licence at https://github.com/cariad/rolumns b
     cook-book/index
     package/index
     genindex
+
+Author
+------
+
+Hello! 👋 I'm **Cariad Eccleston**, and I'm a freelance Amazon Web Services architect, DevOps evangelist, and infrastructure and backend engineer by the beach in the United Kingdom.
+
+You can find me at `cariad.earth <https://cariad.earth>`_, `github.com/cariad <https://github.com/cariad>`_ and `linkedin.com/in/cariad <https://linkedin.com/in/cariad>`_.
