+++ docs/source/package/by_key.rst
@@ -1,7 +1,7 @@
 ByKey
 =====
 
-.. py:module:: rolumns.groups
+.. py:module:: rolumns
     :noindex:
 
 .. autoclass:: ByKey
