+++ docs/source/package/by_user_defined_fields.rst
@@ -1,7 +1,7 @@
 ByUserDefinedFields
 ===================
 
-.. py:module:: rolumns.groups
+.. py:module:: rolumns
     :noindex:
 
 .. autoclass:: ByUserDefinedFields
