--- docs/source/package/groups/index.rst
@@ -1,11 +1,0 @@
-Groups
-======
-
-These pages describe the package's methods of grouping data.
-
-.. toctree::
-    :titlesonly:
-
-    by_key
-    by_path
-    by_user_defined_fields
