+++ docs/source/package/index.rst
@@ -6,8 +6,10 @@ These pages describe the package's modules and classes.
 .. toctree::
     :titlesonly:
 
+    by_key
+    by_path
+    by_user_defined_fields
     columns
     source
-    groups/index
     exceptions
     renderers
