+++ rolumns/__init__.py
@@ -6,7 +6,11 @@ Full documentation is online at https://rolumns.dev.
 
 from importlib.resources import open_text
 
+from rolumns.by_key import ByKey
+from rolumns.by_path import ByPath
+from rolumns.by_user_defined_fields import ByUserDefinedFields, UserDefinedField
 from rolumns.columns import Columns
+from rolumns.group import Group
 from rolumns.source import Source
 from rolumns.translation_state import TranslationState
 
@@ -14,7 +18,12 @@ with open_text(__package__, "VERSION") as t:
     __version__ = t.readline().strip()
 
 __all__ = [
+    "ByKey",
+    "ByPath",
+    "ByUserDefinedFields",
     "Columns",
+    "Group",
     "Source",
     "TranslationState",
+    "UserDefinedField",
 ]
