+++ setup.py
@@ -44,14 +44,12 @@ setup(
         "rolumns",
         "rolumns.enums",
         "rolumns.exceptions",
-        "rolumns.groups",
         "rolumns.renderers",
     ],
     package_data={
         "rolumns": ["py.typed"],
         "rolumns.enums": ["py.typed"],
         "rolumns.exceptions": ["py.typed"],
-        "rolumns.groups": ["py.typed"],
         "rolumns.renderers": ["py.typed"],
     },
     python_requires=">=3.7",
